fix: ensure crop_and_threshold_masks returns all_instances as uint8
The current returned type of all_instances is int32, uint8 is expected.

The current behaviour causes issues in convert_strided_predictions_to_instance_masks as the expected type is not met

+++ research/object_detection/meta_architectures/center_net_meta_arch.py
@@ -2049,7 +2049,7 @@ def crop_and_threshold_masks(elems, input_height, input_width, mask_height=256,
   surface_coords = (
       surface_coords * tf.cast(instance_masks_with_parts, tf.float32))
 
-  return [tf.squeeze(all_instances, axis=3), surface_coords]
+  return [tf.cast(tf.squeeze(all_instances, axis=3), tf.uint8), surface_coords]
 
 
 def gather_surface_coords_for_parts(surface_coords_cropped,
